moved migration scripts out of Util and to more logical places
