+++ app/src/main/java/ie/delilahsthings/soothingloop/CustomSoundsManager.java
@@ -2,12 +2,13 @@ package ie.delilahsthings.soothingloop;
 
 import android.content.Context;
 import android.net.Uri;
+import android.os.Environment;
 
 import java.io.File;
+import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.OutputStream;
 
 public class CustomSoundsManager {
 
@@ -50,5 +51,34 @@ public class CustomSoundsManager {
     {
         public String name;
         public int size;
+    }
+
+    //TODO: If after 2026-04-07 then remove this and anything that depends on it
+    @Deprecated
+    public static boolean migratePre1dot2Noises() {
+        String oldSoundPath = Environment.getDataDirectory().getPath()+"/data/"+StaticContext.getAppContext().getPackageName()+"/custom_sounds/";
+        String newSoundPath = CustomSoundsManager.getSoundPath();
+        File oldSoundDir = new File(oldSoundPath);
+        File newSoundDir = new File(newSoundPath);
+        File inFile, outFile;
+
+        if(oldSoundDir.isDirectory()) {
+            newSoundDir.mkdir();
+
+            try {
+                for (String sound : Util.safe_list(oldSoundDir)) {
+                    inFile = new File(oldSoundPath + sound);
+                    outFile = new File(newSoundPath + sound);
+                    Util.copy(new FileInputStream(inFile), new FileOutputStream(outFile));
+                    inFile.delete();
+                }
+
+                oldSoundDir.delete();
+            } catch (IOException e) {
+                return false;
+            }
+        }
+
+        return true;
     }
 }
