+++ app/src/main/java/ie/delilahsthings/soothingloop/Util.java
@@ -5,14 +5,11 @@ import android.content.Context;
 import android.content.SharedPreferences;
 import android.database.Cursor;
 import android.net.Uri;
-import android.os.Environment;
 import android.provider.OpenableColumns;
 import android.view.View;
 import android.view.ViewGroup;
 
 import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
@@ -67,75 +64,6 @@ public abstract class Util {
             }
         }
         return result;
-    }
-
-    //TODO: If after 2026-04-07 then remove this and anything that depends on it
-    @Deprecated
-    public static boolean migratePre1dot2Noises() {
-        String oldSoundPath = Environment.getDataDirectory().getPath()+"/data/"+StaticContext.getAppContext().getPackageName()+"/custom_sounds/";
-        String newSoundPath = CustomSoundsManager.getSoundPath();
-        File oldSoundDir = new File(oldSoundPath);
-        File newSoundDir = new File(newSoundPath);
-        File inFile, outFile;
-
-        if(oldSoundDir.isDirectory()) {
-            if (!newSoundDir.mkdir()) {
-                return false;
-            }
-
-            try {
-                for (String sound : safe_list(oldSoundDir)) {
-                    inFile = new File(oldSoundPath + sound);
-                    outFile = new File(newSoundPath + sound);
-                    copy(new FileInputStream(inFile), new FileOutputStream(outFile));
-                    inFile.delete();
-                }
-
-                oldSoundDir.delete();
-            } catch (IOException e) {
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    //TODO: If after 2026-04-07 then remove this and anything that depends on it
-    @Deprecated
-    public static boolean migratePre1dot2Profiles() {
-        String oldProfilePath = Environment.getDataDirectory().getPath()+"/data/"+StaticContext.getAppContext().getPackageName()+"/shared_prefs/";
-        String newProfilePath = ProfileManager.getProfilePath();
-        File oldProfileDir = new File(oldProfilePath);
-        File newProfileDir = new File(newProfilePath);
-        File inFile, outFile;
-        newProfileDir.mkdirs();
-
-        if(!oldProfileDir.isDirectory()) {
-            return true;
-        }
-
-        try {
-            inFile = new File(oldProfilePath + "default.xml");
-            outFile = ProfileManager.getDefaultProfileFile();
-
-            if(inFile.exists()) {
-                copy(new FileInputStream(inFile), new FileOutputStream(outFile));
-                inFile.delete();
-            }
-
-            for (String profile : safe_list(oldProfileDir)) {
-                if(profile.startsWith("profile-")) {
-                    inFile = new File(oldProfilePath + profile);
-                    outFile = new File(newProfilePath + profile.substring(8));
-                    copy(new FileInputStream(inFile), new FileOutputStream(outFile));
-                    inFile.delete();
-                }
-            }
-        } catch (IOException e) {
-            return false;
-        }
-
-        return true;
     }
 
     /**
